recover OK. (word_spacing old version) (sentences=100)

+++ nlp4kor/ffnn/word_spacing.sh
@@ -2,27 +2,23 @@
 PROGRAM="word_spacing.py"
 PROJECT_DIR=${HOME}'/workspace/nlp4kor'
 
-echo "[Usage] $0 n_train ngram"
-echo "[Usage] $0 10000 6"
-echo "[Usage] $0 10000 4"
-
-#if [ "$#" -ne 2 ]; then
-#    echo "[Usage] $0 n_train ngram"
-#    echo "[Usage] $0 10000 6"
-#    echo "[Usage] $0 10000 4"
-#    exit
-#fi
+if [ "$#" -ne 3 ]; then
+    echo "[Usage] $0 n_train left_gram right_gram"
+    echo "[Usage] $0 100 2 2"
+    exit
+fi
 
 n_train=$1
-ngram=$2
+left_gram=$2
+right_gram=$3
 
 #git --work-tree=${PROJECT_DIR} --git-dir=${PROJECT_DIR}/.git pull
 
 echo "pkill -f ${PROGRAM}"
 pkill -f ${PROGRAM}
 
-echo "rm -f logs/${PROGRAM}.$1.$2.*"
-rm -f logs/${PROGRAM}.$1.$2.*
+echo "rm -f logs/${PROGRAM}.$1.$2.$3.*"
+rm -f logs/${PROGRAM}.$1.$2.$3.*
 
 echo "python3 ./${PROGRAM} $n_train $ngram >/dev/null 2>&1 &"
 nohup python3 ./${PROGRAM} ${n_train} ${ngram} >/dev/null 2>&1 &
